requirements.txt
django>=3.0.2
django-reversion>=3.0.5